+++ src/DataFactory/DataFactoryV2.Test/ScenarioTests/RunTests.cs
@@ -30,7 +30,7 @@ namespace Microsoft.Azure.Commands.DataFactoryV2.Test
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
-        public void TestRunV2()
+        public void TestPipelineRunV2()
         {
             RunPowerShellTest(_logger, "Test-Run");
         }