+++ src/PolicyInsights/PolicyInsights/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Improved error messages
 
 ## Version 1.2.0
 * Support evaluating compliance prior to determining what resource to remediate