+++ src/PolicyInsights/PolicyInsights/ChangeLog.md
@@ -18,7 +18,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-* Improve error messages
+
+## Version 1.2.1
+* Improved error messages
 
 ## Version 1.2.0
 * Support evaluating compliance prior to determining what resource to remediate